requirements.txt
Django==3.2.6
djangorestframework==3.12.4
PyMySQL==1.0.2
mysqlclient==2.0.3
requests~=2.26.0